producer/requirements.txt
kafka-python
praw
dotenv
retry